requirements.txt
llm-guard>=0.3.0
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
colorlog>=6.7.0
